refactor: remove `Predicate` usage from `QueryNamespace` (#8468)
For #8097.
