+++ iox_query/src/lib.rs
@@ -21,7 +21,11 @@ use arrow::{
 };
 use async_trait::async_trait;
 use data_types::{ChunkId, ChunkOrder, TransitionPartitionId};
-use datafusion::{error::DataFusionError, physical_plan::Statistics, prelude::SessionContext};
+use datafusion::{
+    error::DataFusionError,
+    physical_plan::Statistics,
+    prelude::{Expr, SessionContext},
+};
 use exec::IOxSessionContext;
 use once_cell::sync::Lazy;
 use parquet_file::storage::ParquetExecInput;
@@ -146,9 +150,9 @@ pub type QueryText = Box<dyn std::fmt::Display + Send + Sync>;
 #[async_trait]
 pub trait QueryNamespace: QueryNamespaceMeta + Debug + Send + Sync {
     /// Returns a set of chunks within the partition with data that may match the provided
-    /// predicate.
+    /// filter expression.
     ///
-    /// If possible, chunks which have no rows that can possibly match the predicate may be omitted.
+    /// If possible, chunks which have no rows that can possibly match the filter may be omitted.
     ///
     /// If projection is `None`, returned chunks will include all columns of its original data.
     /// Otherwise, returned chunks will include PK columns (tags and time) and columns specified in
@@ -157,7 +161,7 @@ pub trait QueryNamespace: QueryNamespaceMeta + Debug + Send + Sync {
     async fn chunks(
         &self,
         table_name: &str,
-        predicate: &Predicate,
+        filters: &[Expr],
         projection: Option<&Vec<usize>>,
         ctx: IOxSessionContext,
     ) -> Result<Vec<Arc<dyn QueryChunk>>, DataFusionError>;
